Before experimenting with timestamps

+++ app/src/main/java/com/example/accelerometer/AccelerationDataDbHelper.java
@@ -19,7 +19,6 @@ public class AccelerationDataDbHelper extends SQLiteOpenHelper {
     private static final String DATABASE_NAME = "accelerationData.db";
     private static final int DATABASE_VERSION = 1;
     public static final String TABLE_NAME = "acceleration_data";
-    public static final String COLUMN_ID = "id";
     public static final String COLUMN_TOTAL_ACCELERATION = "acceleration";
     public static final String COLUMN_X_ACCELERATION = "x";
     public static final String COLUMN_Y_ACCELERATION = "y";
@@ -34,7 +33,7 @@ public class AccelerationDataDbHelper extends SQLiteOpenHelper {
 
     @Override
     public void onCreate(SQLiteDatabase db) {
-        String CREATE_TABLE = "CREATE TABLE acceleration_data (" +
+        String CREATE_TABLE = "CREATE TABLE IF NOT EXISTS acceleration_data (" +
                 "id INTEGER PRIMARY KEY AUTOINCREMENT," +
                 " acceleration REAL," +
                 " x REAL," +
@@ -45,28 +44,17 @@ public class AccelerationDataDbHelper extends SQLiteOpenHelper {
     }
 
 
-    /*
-    onUpgrade() is called (you do not call it yourself) when version of your DB changed which means underlying table structure changed etc.
-
-    In general it means that OS is telling you "hey, you asked for database structure version 10 but
-    I found we got something older here, so this is you chance to fix that before you start using database (and potentially crash due to structure mismatch)".
-
-    In that method you should do all that is necessary to, well.. upgrade structure of your old database to structure matching current version requirements
-    like adding/dropping columns, converting row contents or even dropping old db completely and create it from scratch
-    For Android it does not matter what you do here - it's just a sort of emergency callback for your code to do the necessary job (if any).
-    You need to be aware that users may not update frequently so you have to always handle upgrade from version X to Y knowing that X may not be equal to i.e. (Y-1).
-    https://stackoverflow.com/questions/13159210/onupgrade-sqlite-database-in-android
-
-    TL;DR IT GETS CALLED WHENEVER YOU CHANGE DB VERSION
-    */
-
 
     @Override
-    public void onUpgrade(SQLiteDatabase db, int i, int i1) {
+    public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
+
+        if(oldVersion == newVersion){
+            return;
+        }
         db.execSQL("DROP TABLE IF EXISTS " + TABLE_NAME);
         onCreate(db);
-
     }
+
 
     public void addAccelerationData(float acceleration, float x, float y, float z, long timestamp) {
         SQLiteDatabase db = this.getWritableDatabase();
@@ -83,7 +71,7 @@ public class AccelerationDataDbHelper extends SQLiteOpenHelper {
     }
 
 
-    // Get the highest acceleration within a time range
+
     public double getHighestAcceleration() {
 
         SQLiteDatabase db = this.getReadableDatabase();
@@ -100,7 +88,6 @@ public class AccelerationDataDbHelper extends SQLiteOpenHelper {
     }
 
 
-    // Get the average acceleration within a time range
     public double getAverageAcceleration() {
         SQLiteDatabase db = this.getReadableDatabase();
         String query = "SELECT AVG(acceleration) FROM " + TABLE_NAME + " WHERE timestamp BETWEEN ? AND ?";
@@ -115,7 +102,7 @@ public class AccelerationDataDbHelper extends SQLiteOpenHelper {
         return averageAcceleration;
     }
 
-    // Get the time spent above a certain speed limit within a time range
+
     public double getTimeSpentAboveLimit() {
         SQLiteDatabase db = this.getReadableDatabase();
         String query = "SELECT SUM((timestamp - prev_timestamp) / 1000.0) " +
@@ -133,77 +120,63 @@ public class AccelerationDataDbHelper extends SQLiteOpenHelper {
         return timeSpentAboveLimit;
     }
 
-    // Get the percentage of time spent above a certain acceleration threshold within a time range
+
+
     public double getPercentageAboveThreshold() {
         double timeAboveThreshold = getTimeSpentAboveLimit();
         double totalTime = (endTimestamp - startTimestamp) / 1000.0;
         return (timeAboveThreshold / totalTime) * 100.0;
     }
 
-    // Get the count of aggressive braking events within a time range
+
     public int getAggressiveAccelerationCount() {
         SQLiteDatabase db = this.getReadableDatabase();
-
-        // Query to get all acceleration values below -2.98 within the time range
         String query = "SELECT acceleration FROM " + TABLE_NAME + " WHERE timestamp BETWEEN ? AND ? AND acceleration";
-
-        // Execute the query
         String[] args = { String.valueOf(startTimestamp), String.valueOf(endTimestamp) };
         Cursor cursor = db.rawQuery(query, args);
 
-        // Count the number of times the acceleration value goes below -2.98
         int aggressiveAccelerationCount = 0;
         boolean wasAboveThreshold = false;
         if (cursor != null && cursor.moveToFirst()) {
             do {
                 double acceleration = cursor.getDouble(0);
                 if (!wasAboveThreshold && acceleration >= 2.98) {
-                    // The acceleration value has just gone above 2.98, so increment the count
                     aggressiveAccelerationCount++;
                     wasAboveThreshold = true;
                 } else if (wasAboveThreshold && acceleration < 2.98) {
-                    // The acceleration value has just gone above -2.98 after being below it, so reset the flag
                     wasAboveThreshold = false;
                 }
             } while (cursor.moveToNext());
 
             cursor.close();
         }
-
         db.close();
         return aggressiveAccelerationCount;
     }
 
-    // Get the count of aggressive acceleration events within a time range
+
     public int getAggressiveBrakingCount() {
+
         SQLiteDatabase db = this.getReadableDatabase();
-
-        // Query to get all acceleration values below -2.98 within the time range
         String query = "SELECT acceleration FROM " + TABLE_NAME + " WHERE timestamp BETWEEN ? AND ?";
-
-        // Execute the query
         String[] args = { String.valueOf(startTimestamp), String.valueOf(endTimestamp) };
         Cursor cursor = db.rawQuery(query, args);
 
-        // Count the number of times the acceleration value goes below -2.98
         int aggressiveBrakingCount = 0;
         boolean wasBelowThreshold = false;
         if (cursor != null && cursor.moveToFirst()) {
             do {
                 double acceleration = cursor.getDouble(0);
                 if (!wasBelowThreshold && acceleration < -2.98) {
-                    // The deceleration value has just gone below -2.98, so increment the count
                     aggressiveBrakingCount++;
                     wasBelowThreshold = true;
                 } else if (wasBelowThreshold && acceleration >= -2.98) {
-                    // The deceleration value has just gone above -2.98 after being below it, so reset the flag
                     wasBelowThreshold = false;
                 }
             } while (cursor.moveToNext());
 
             cursor.close();
         }
-
         db.close();
         return aggressiveBrakingCount;
     }
@@ -303,12 +276,14 @@ public class AccelerationDataDbHelper extends SQLiteOpenHelper {
     public void getTimestamps(int selection) {
         long[] timestamps = new long[2];
 
+        Calendar calendar = Calendar.getInstance();
+        calendar.set(Calendar.HOUR_OF_DAY, 0);
+        calendar.set(Calendar.MINUTE, 0);
+        calendar.set(Calendar.SECOND, 0);
+
         switch (selection) {
             case 1:
-                Calendar calendar = Calendar.getInstance();
-                calendar.set(Calendar.HOUR_OF_DAY, 0);
-                calendar.set(Calendar.MINUTE, 0);
-                calendar.set(Calendar.SECOND, 0);
+
                 timestamps[0] = calendar.getTimeInMillis();
                 timestamps[1] = System.currentTimeMillis();
                 break;
